refactor: modularize db class
- move profile subclass to a different file
